extract index methods
