+++ packages/mongo/collection/collection.js
@@ -1,6 +1,7 @@
 import { normalizeProjection } from "../mongo_utils";
 import { AsyncMethods } from './methods_async';
 import { SyncMethods } from './methods_sync';
+import { IndexMethods } from './methods_index';
 
 /**
  * @summary Namespace for MongoDB-related items
@@ -481,10 +482,6 @@ Object.assign(Mongo.Collection.prototype, {
       };
     }
   },
-
-
-
-
 });
 
 Object.assign(Mongo.Collection, {
@@ -542,111 +539,14 @@ Object.assign(Mongo.Collection, {
   },
 });
 
-Object.assign(Mongo.Collection.prototype, SyncMethods);
-Object.assign(Mongo.Collection.prototype, AsyncMethods);
+Object.assign(Mongo.Collection.prototype, SyncMethods, AsyncMethods, IndexMethods);
 
 Object.assign(Mongo.Collection.prototype, {
-
   // Determine if this collection is simply a minimongo representation of a real
   // database on another server
   _isRemoteCollection() {
     // XXX see #MeteorServerNull
     return this._connection && this._connection !== Meteor.server;
-  },
-
-
-
-
-
-
-  // We'll actually design an index API later. For now, we just pass through to
-  // Mongo's, but make it synchronous.
-  /**
-   * @summary Asynchronously creates the specified index on the collection.
-   * @locus server
-   * @method ensureIndexAsync
-   * @deprecated in 3.0
-   * @memberof Mongo.Collection
-   * @instance
-   * @param {Object} index A document that contains the field and value pairs where the field is the index key and the value describes the type of index for that field. For an ascending index on a field, specify a value of `1`; for descending index, specify a value of `-1`. Use `text` for text indexes.
-   * @param {Object} [options] All options are listed in [MongoDB documentation](https://docs.mongodb.com/manual/reference/method/db.collection.createIndex/#options)
-   * @param {String} options.name Name of the index
-   * @param {Boolean} options.unique Define that the index values must be unique, more at [MongoDB documentation](https://docs.mongodb.com/manual/core/index-unique/)
-   * @param {Boolean} options.sparse Define that the index is sparse, more at [MongoDB documentation](https://docs.mongodb.com/manual/core/index-sparse/)
-   */
-  async ensureIndexAsync(index, options) {
-    var self = this;
-    if (!self._collection.ensureIndexAsync || !self._collection.createIndexAsync)
-      throw new Error('Can only call createIndexAsync on server collections');
-    if (self._collection.createIndexAsync) {
-      await self._collection.createIndexAsync(index, options);
-    } else {
-      import { Log } from 'meteor/logging';
-
-      Log.debug(`ensureIndexAsync has been deprecated, please use the new 'createIndexAsync' instead${ options?.name ? `, index name: ${ options.name }` : `, index: ${ JSON.stringify(index) }` }`)
-      await self._collection.ensureIndexAsync(index, options);
-    }
-  },
-
-  /**
-   * @summary Asynchronously creates the specified index on the collection.
-   * @locus server
-   * @method createIndexAsync
-   * @memberof Mongo.Collection
-   * @instance
-   * @param {Object} index A document that contains the field and value pairs where the field is the index key and the value describes the type of index for that field. For an ascending index on a field, specify a value of `1`; for descending index, specify a value of `-1`. Use `text` for text indexes.
-   * @param {Object} [options] All options are listed in [MongoDB documentation](https://docs.mongodb.com/manual/reference/method/db.collection.createIndex/#options)
-   * @param {String} options.name Name of the index
-   * @param {Boolean} options.unique Define that the index values must be unique, more at [MongoDB documentation](https://docs.mongodb.com/manual/core/index-unique/)
-   * @param {Boolean} options.sparse Define that the index is sparse, more at [MongoDB documentation](https://docs.mongodb.com/manual/core/index-sparse/)
-   */
-  async createIndexAsync(index, options) {
-    var self = this;
-    if (!self._collection.createIndexAsync)
-      throw new Error('Can only call createIndexAsync on server collections');
-
-    try {
-      await self._collection.createIndexAsync(index, options);
-    } catch (e) {
-      if (
-        e.message.includes(
-          'An equivalent index already exists with the same name but different options.'
-        ) &&
-        Meteor.settings?.packages?.mongo?.reCreateIndexOnOptionMismatch
-      ) {
-        import { Log } from 'meteor/logging';
-
-        Log.info(`Re-creating index ${ index } for ${ self._name } due to options mismatch.`);
-        await self._collection.dropIndexAsync(index);
-        await self._collection.createIndexAsync(index, options);
-      } else {
-        console.error(e);
-        throw new Meteor.Error(`An error occurred when creating an index for collection "${ self._name }: ${ e.message }`);
-      }
-    }
-  },
-
-  /**
-   * @summary Asynchronously creates the specified index on the collection.
-   * @locus server
-   * @method createIndex
-   * @memberof Mongo.Collection
-   * @instance
-   * @param {Object} index A document that contains the field and value pairs where the field is the index key and the value describes the type of index for that field. For an ascending index on a field, specify a value of `1`; for descending index, specify a value of `-1`. Use `text` for text indexes.
-   * @param {Object} [options] All options are listed in [MongoDB documentation](https://docs.mongodb.com/manual/reference/method/db.collection.createIndex/#options)
-   * @param {String} options.name Name of the index
-   * @param {Boolean} options.unique Define that the index values must be unique, more at [MongoDB documentation](https://docs.mongodb.com/manual/core/index-unique/)
-   * @param {Boolean} options.sparse Define that the index is sparse, more at [MongoDB documentation](https://docs.mongodb.com/manual/core/index-sparse/)
-   */
-  createIndex(index, options){
-    return this.createIndexAsync(index, options);
-  },
-
-  async dropIndexAsync(index) {
-    var self = this;
-    if (!self._collection.dropIndexAsync)
-      throw new Error('Can only call dropIndexAsync on server collections');
-    await self._collection.dropIndexAsync(index);
   },
 
   async dropCollectionAsync() {
